Minor fixes to eliminate warnings when compiling with gcc 5.3 and CXXFLAGS="-Wall -Wextra -Wfloat-conversion -Wno-unused-value -Wno-unused-local-typedefs -Wno-unused-parameter"

+++ mp-algorithms/stateslist.cpp
@@ -200,7 +200,7 @@ void StatesList::AppendToken(char const* s)
    {
       if (FinalStates != 0)
       {
-	 I.NumStates = InitialStates + i*double((FinalStates-InitialStates)/double(NumSweeps-1));
+	 I.NumStates = InitialStates + int(i*double((FinalStates-InitialStates)/double(NumSweeps-1)));
       }
       else if (Increment != 0)
       {
